requirements.txt
Django<2.3
factory_boy==2.11.1
beautifulsoup4==4.7.1
